backend/requirements.txt
# FastAPI and related packages
fastapi==0.103.1
uvicorn==0.23.2
gunicorn==21.2.0
pydantic==2.3.0

# CORS support
starlette==0.27.0

# Data processing
numpy==1.24.3

# Google AI for embeddings
google-generativeai==0.3.0
grpcio==1.54.0
grpcio-status==1.54.0

# Environment variables
python-dotenv==1.0.0

# Type hints
typing-extensions==4.7.1

# Add these for robustness
protobuf==4.24.4
six==1.16.0